Introduces a service layer for data access
This commit introduces a service layer to abstract data access logic.

It creates a `BackendService` to handle data fetching from the repository and injects it into the `DashboardBloc`.
This improves separation of concerns and allows for easier testing and future data source modifications.

The `DashboardBloc` now retrieves data from the `BackendService` instead of directly from the `MockRepository`.
The main application now provides the BackendService using RepositoryProvider

This approach makes the application more modular and maintainable.